# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Compte API",
        "description": "API Documentation pour la gestion des comptes utilisateurs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Obtenir tous les messages",
                "operationId": "9d5effa2cb648df85be0ea9741ffb457",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "chemin/vers/votre/fichier.yaml#/components/schemas/Message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messageStore": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Ajouter un nouveau message",
                "operationId": "4d6b3cf80959a63587ba2d1f0be1e89e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Ceci est un message"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "exemple@gmail.com"
                                    },
                                    "nomComplet": {
                                        "type": "string",
                                        "example": "Adama Gueye"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message ajouté avec succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messageShow{id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Obtenir un message spécifique par ID",
                "operationId": "1e1f00bbe03ab2ea3b600e526029e5ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du message",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Message non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messageDestroy{id}": {
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Supprimer un message",
                "operationId": "6c81c8c51a6358f01e4831c1e6cdb005",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du message",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message supprimé avec succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Message non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/newsLetters": {
            "get": {
                "tags": [
                    "Newsletters"
                ],
                "summary": "Obtenir toutes les newsletters",
                "operationId": "3b269ae8bdcc95dcda5cc625ad587b26",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NewsLetter"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/newsLetterStore": {
            "post": {
                "tags": [
                    "Newsletters"
                ],
                "summary": "S'abonner à la newsletter",
                "operationId": "012572400c30cdd0ea79e772aab39a60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "exemple@exemple.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscription réussie"
                    },
                    "422": {
                        "description": "Email déjà inscrit"
                    },
                    "400": {
                        "description": "Entrée non valide"
                    }
                }
            }
        },
        "/api/newsLetterDestroy{id}": {
            "delete": {
                "tags": [
                    "Newsletters"
                ],
                "summary": "Se désabonner de la newsletter",
                "operationId": "78f9a55b4c56263ef8490100ddc7849d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'abonnement à la newsletter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Désinscription réussie"
                    },
                    "404": {
                        "description": "Abonnement non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}